Updated Transaction types
