+++ Sources/Starknet/Data/Transaction/Transaction.swift
@@ -1,52 +1,5 @@
 import Foundation
 
-public struct StarknetSequencerInvokeTransaction: StarknetSequencerTransaction, Equatable {
-    public let type: StarknetTransactionType = .invoke
-
-    public let senderAddress: Felt
-
-    public let calldata: StarknetCalldata
-
-    public let signature: StarknetSignature
-
-    public let maxFee: Felt
-
-    public let nonce: Felt
-
-    public let version: Felt
-
-    enum CodingKeys: String, CodingKey {
-        case type
-        case version
-        case senderAddress = "sender_address"
-        case calldata
-        case signature
-        case maxFee = "max_fee"
-        case nonce
-    }
-
-    public init(senderAddress: Felt, calldata: StarknetCalldata, signature: StarknetSignature, maxFee: Felt, nonce: Felt, version: Felt) {
-        self.senderAddress = senderAddress
-        self.calldata = calldata
-        self.signature = signature
-        self.maxFee = maxFee
-        self.nonce = nonce
-        self.version = version
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-
-        self.senderAddress = try container.decode(Felt.self, forKey: .senderAddress)
-        self.calldata = try container.decode(StarknetCalldata.self, forKey: .calldata)
-        self.signature = try container.decode(StarknetSignature.self, forKey: .signature)
-        self.maxFee = try container.decode(Felt.self, forKey: .maxFee)
-        self.nonce = try container.decode(Felt.self, forKey: .nonce)
-        self.version = try container.decode(Felt.self, forKey: .version)
-
-        try verifyTransactionIdentifiers(container: container, codingKeysType: CodingKeys.self)
-    }
-}
 
 public struct StarknetInvokeTransaction: StarknetTransaction, Equatable {
     public let type: StarknetTransactionType = .invoke
@@ -169,59 +122,6 @@ public struct StarknetInvokeTransactionV0: StarknetTransaction, Equatable {
     }
 }
 
-
-public struct StarknetSequencerDeployAccountTransaction: StarknetSequencerTransaction, Equatable {
-    public let type: StarknetTransactionType = .deployAccount
-
-    public let version: Felt
-
-    public let signature: StarknetSignature
-
-    public let maxFee: Felt
-
-    public let nonce: Felt
-
-    public let contractAddressSalt: Felt
-
-    public let constructorCalldata: StarknetCalldata
-
-    public let classHash: Felt
-
-    public init(signature: StarknetSignature, maxFee: Felt, nonce: Felt, contractAddressSalt: Felt, constructorCalldata: StarknetCalldata, classHash: Felt, version: Felt) {
-        self.signature = signature
-        self.maxFee = maxFee
-        self.nonce = nonce
-        self.contractAddressSalt = contractAddressSalt
-        self.constructorCalldata = constructorCalldata
-        self.classHash = classHash
-        self.version = version
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.signature = try container.decode(StarknetSignature.self, forKey: .signature)
-        self.maxFee = try container.decode(Felt.self, forKey: .maxFee)
-        self.nonce = try container.decode(Felt.self, forKey: .nonce)
-        self.contractAddressSalt = try container.decode(Felt.self, forKey: .contractAddressSalt)
-        self.constructorCalldata = try container.decode(StarknetCalldata.self, forKey: .constructorCalldata)
-        self.classHash = try container.decode(Felt.self, forKey: .classHash)
-        self.version = try container.decode(Felt.self, forKey: .version)
-
-        try verifyTransactionIdentifiers(container: container, codingKeysType: CodingKeys.self)
-    }
-
-    enum CodingKeys: String, CodingKey {
-        case type
-        case version
-        case signature
-        case maxFee = "max_fee"
-        case nonce
-        case contractAddressSalt = "contract_address_salt"
-        case constructorCalldata = "constructor_calldata"
-        case classHash = "class_hash"
-    }
-}
-
 public struct StarknetDeployAccountTransaction: StarknetTransaction, Equatable {
     public let type: StarknetTransactionType = .deployAccount
 
@@ -292,7 +192,7 @@ public struct StarknetDeployAccountTransaction: StarknetTransaction, Equatable {
     }
 }
 
-public struct StarknetL1HandlerTransaction: StarknetTransaction {
+public struct StarknetL1HandlerTransaction: StarknetTransaction, Equatable {
     public let type: StarknetTransactionType = .l1Handler
 
     public let version: Felt
